+++ Assets/Framework/Scripts/Runtime/Modules/Nova/Nova.Visitors.cs
@@ -24,7 +24,7 @@ namespace NovaFramework.Runtime
         /// <summary>
         /// Nova 版本号。
         /// </summary>
-        public const string Version = "0.5.40";
+        public const string Version = "0.5.41";
 
         /// <summary>
         /// Nova 自身引用。